campus-access-control/requirements_notebook.txt
# Core Computer Vision
opencv-python==4.12.0.86
opencv-contrib-python==4.12.0.86

# OCR
easyocr==1.7.2

# Deep Learning - PyTorch
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# YOLO
ultralytics==8.3.60

# Face Recognition
deepface==0.0.93
tf-keras==2.18.0
tensorflow==2.18.0

# Database
psycopg2-binary==2.9.10
python-dotenv==1.1.1

# Data Processing
numpy<2.0
pandas==2.2.3
matplotlib==3.9.4

# Jupyter Support
ipywidgets==8.1.5
ipython==8.29.0
jupyter==1.1.1

# Utilities
tqdm==4.67.1
Pillow==11.0.0
scipy==1.15.1

# Additional Dependencies
requests==2.32.3
pyyaml==6.0.2
